% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRV_field.R
\docType{data}
\name{extract_HRV_field}
\alias{extract_HRV_field}
\title{Extract HRV fields}
\format{
An object of class \code{list} of length 39.
}
\usage{
extract_HRV_field
}
\description{
A Helper function to extract various HRV fields
from a character vector of LabChart HRV report.
}
\keyword{datasets}
