% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HRV_vars_desc}
\alias{HRV_vars_desc}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 2 columns:
\describe{
\item{variable}{a variable names}
\item{description}{a description of each variables}
...
}
}
\usage{
HRV_vars_desc
}
\description{
A dataset containing variable description of HRV report when imported into \code{R} via
\code{\link[=read_HRV_reports]{read_HRV_reports()}} or \code{\link[=parse_HRV_reports]{parse_HRV_reports()}}
}
\keyword{datasets}
