% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HRV_vars_domain}
\alias{HRV_vars_domain}
\title{HRV Time & Frequency Domain Variable}
\format{
A list with 2 elements:
\describe{
\item{time}{contain character vector of time-domain variables}
\item{freq}{contain character vector of frequency-domain variables}
}
}
\usage{
HRV_vars_domain
}
\description{
A list for quickly select HRV variables in the category of time or frequency domain.
It is intended to be used with the resulting tibble of \code{\link[=read_HRV_reports]{read_HRV_reports()}} or \code{\link[=parse_HRV_reports]{parse_HRV_reports()}}.
}
\keyword{datasets}
