% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{labChartHRV_example}
\alias{labChartHRV_example}
\title{Get path to labChartHRV example}
\usage{
labChartHRV_example(path = NULL)
}
\arguments{
\item{path}{Path to example file or folder. If \code{NULL}, the example files or folders will be listed.}
}
\description{
labChartHRV comes bundled with a number of sample files in its \code{inst/extdata}
directory. This function make them easy to access
}
\examples{
library(labChartHRV)
labChartHRV_example()
labChartHRV_example("HRV")
}
