% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRV_reports.R
\name{parse_HRV_reports}
\alias{parse_HRV_reports}
\title{Parse HRV Report file}
\usage{
parse_HRV_reports(x, format_cols = TRUE)
}
\arguments{
\item{x}{(Character) Character vectors of HRV report, each elements
corresponding to one report.}

\item{format_cols}{If \code{TRUE} format appropriate output variables as factor, integer, or double. Compute \code{NN50_percent} (pNN50) from \code{NN50_count/Normals_count}}
}
\value{
A tibble
}
\description{
Parse HRV Report file from a character vector to a tibble.
}
\examples{
NULL
}
