% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRV_reports.R
\name{read_HRV_reports}
\alias{read_HRV_reports}
\title{Read LabChart's HRV Report File}
\usage{
read_HRV_reports(file, format_cols = TRUE, encoding = "UTF-16LE", ...)
}
\arguments{
\item{file}{(Character) Path to LabChart's HRV report (\code{.txt}) files to be read.
Path can point to an individual text file (\code{.txt}) or a folder which contains those files.}

\item{format_cols}{(Logical) If \code{TRUE} formats appropriate output variables as factor, integer, or double, and compute \code{NN50_percent} (pNN50) from \code{NN50_count/Normals_count}}

\item{encoding}{(Character) Text encoding of the HRV report files.
Default value is set to "UTF-16LE" as it works well with my testing data.
If the result seems abnormal, try different text encoding (visit \href{https://readtext.quanteda.io/index.html}{readtext manual} to see all available options)}

\item{...}{extra arguments passed to \code{\link[readtext:readtext]{readtext::readtext()}}}
}
\value{
A tibble of HRV report (one row per subject)
}
\description{
Read and transform LabChart's HRV report file (\code{.txt}) to a tidy \href{https://tibble.tidyverse.org}{tibble}
with one row per file.
}
\details{
\code{\link[=read_HRV_reports]{read_HRV_reports()}} reads contents of LabChart's HRV report from text file (\code{.txt})
into character vector using \code{\link[readtext:readtext]{readtext::readtext()}} function with default text encoding as "UTF-16LE".
Then, each key-value fields of the character vector will be extracted and transform into a \href{https://tibble.tidyverse.org}{tibble}
by regular expression pattern matching as in \code{\link[=parse_HRV_reports]{parse_HRV_reports()}}.

The result will be a tibble with column names corresponding to "key" in the input file(s),
and values in each rows mapped to "value" from the input file(s).
}
\examples{
# Path to `HRV/` folder containing example HRV report files (`.txt`)
path <- sipsANS_example("HRV")
# Read into tibble
read_HRV_reports(path)
}
