% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{sipsANS_example}
\alias{sipsANS_example}
\title{Get path to sipsANS example}
\usage{
sipsANS_example(path = NULL)
}
\arguments{
\item{path}{Path to example file or folder. If \code{NULL}, the example files or folders will be listed.}
}
\description{
sipsANS comes bundled with a number of sample files in its \code{inst/extdata}
directory. This function make them easy to access
}
\examples{
library(sipsANS)
sipsANS_example()
sipsANS_example("HRV")
}
